Unify slider control between Volume Panel and Volume Dialog
The new Volume Dialog and Volume Panel will both use the same Slider
control, that handles haptics, a11y and animated snapping. This change
mostly copies the VolumeSlider used in the Volume Panel and adapts it
for the Volume Dialog.

Flag: com.android.systemui.volume_redesign
Fixes: 393133850
Fixes: 395578673
Test: atest VolumeDialogScreenshotTest
Test: atest VolumePanelScreenshotTest
Test: manual on foldable

+++ packages/SystemUI/src/com/android/systemui/volume/panel/component/volume/slider/ui/viewmodel/SliderState.kt
@@ -36,7 +36,7 @@ sealed interface SliderState {
      * A11y slider controls works by adjusting one step up or down. The default slider step isn't
      * enough to trigger rounding to the correct value.
      */
-    val a11yStep: Int
+    val a11yStep: Float
     val a11yClickDescription: String?
     val a11yStateDescription: String?
     val disabledMessage: String?
@@ -49,7 +49,7 @@ sealed interface SliderState {
         override val icon: Icon? = null
         override val label: String = ""
         override val disabledMessage: String? = null
-        override val a11yStep: Int = 0
+        override val a11yStep: Float = 0f
         override val a11yClickDescription: String? = null
         override val a11yStateDescription: String? = null
         override val isEnabled: Boolean = true
